+++ graspit-reporters/packages/graspit-commons/feed.cjs
@@ -5,4 +5,4 @@ const { platform } = require('os');
 
 const args = process.argv.slice(2);
 const executable = platform() === 'win32' ? 'graspit.exe' : 'graspit';
-spawnSync(join(__dirname, executable), args, { stdio: 'inherit', env: process.env });
+spawnSync(join(__dirname, 'bin', executable), args, { stdio: 'inherit', env: process.env });